pyproject.toml
[project]
name = "search-daemon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "sentence-transformers>=3.0",
    "chromadb>=0.5",
    "watchdog>=4.0",
    "pyyaml>=6.0",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "python-pptx>=0.6",
    "openpyxl>=3.1",
    "rumps>=0.4",
]

[project.scripts]
search-daemon = "search_daemon.__main__:main"
search-menu-bar = "search_daemon.menu_bar:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
